requirements.txt
Django >= 4
celery[redis]
djangorestframework
django-cacheops
django-cors-headers
drf-spectacular
paramiko
pymysql
redis
